+++ docs/src/Fulma/Components/Dropdown/View.fs
@@ -12,7 +12,7 @@ open Fulma.Extra.FontAwesome
 let basic =
     Dropdown.dropdown [ Dropdown.isHoverable ]
         [ div [ ]
-            [ Button.button_div [ ]
+            [ Button.button_a [ ]
                 [ span [ ]
                     [ str "Dropdown" ]
                   Icon.faIcon [ Icon.isSmall ] [ Fa.icon Fa.I.AngleDown ] ] ]