+++ docs/src/Fulma/Components/Panel/State.fs
@@ -31,10 +31,10 @@ let iconCode =
         [ input [ Type "checkbox" ]
           str "I am a checkbox" ]
       Panel.block [ ]
-        [ Button.button_div [ Button.isPrimary
-                              Button.isOutlined
-                              Button.isFullWidth ]
-                            [ str "Reset" ] ] ]
+        [ Button.button_a [ Button.isPrimary
+                            Button.isOutlined
+                            Button.isFullWidth ]
+                          [ str "Reset" ] ] ]
 ```
     """