+++ docs/src/Fulma/Elements/Button/State.fs
@@ -6,27 +6,27 @@ open Types
 let colorCode =
     """
 ```fsharp
-    Button.button_div [ ] [ str "Button" ]
-    Button.button_div [ Button.isWhite ] [ str "White" ]
-    Button.button_div [ Button.isLight ] [ str "Light" ]
-    Button.button_div [ Button.isDark ] [ str "Dark" ]
-    Button.button_div [ Button.isBlack ] [ str "Black" ]
-    Button.button_div [ Button.isLink ] [ str "Link" ]
-    Button.button_div [ Button.isPrimary ] [ str "Primary" ]
-    Button.button_div [ Button.isInfo ] [ str "Info" ]
-    Button.button_div [ Button.isSuccess ] [ str "Success" ]
-    Button.button_div [ Button.isWarning ] [ str "Warning" ]
-    Button.button_div [ Button.isDanger ] [ str "Danger" ]
+    Button.button_a [ ] [ str "Button" ]
+    Button.button_a [ Button.isWhite ] [ str "White" ]
+    Button.button_a [ Button.isLight ] [ str "Light" ]
+    Button.button_a [ Button.isDark ] [ str "Dark" ]
+    Button.button_a [ Button.isBlack ] [ str "Black" ]
+    Button.button_a [ Button.isLink ] [ str "Link" ]
+    Button.button_a [ Button.isPrimary ] [ str "Primary" ]
+    Button.button_a [ Button.isInfo ] [ str "Info" ]
+    Button.button_a [ Button.isSuccess ] [ str "Success" ]
+    Button.button_a [ Button.isWarning ] [ str "Warning" ]
+    Button.button_a [ Button.isDanger ] [ str "Danger" ]
 ```
     """
 
 let sizeCode =
     """
 ```fsharp
-    Button.button_div [ Button.isSmall ] [ str "Small" ]
-    Button.button_div [ ] [ str "Normal" ]
-    Button.button_div [ Button.isMedium ] [ str "Medium" ]
-    Button.button_div [ Button.isLarge ] [ str "Large" ]
+    Button.button_a [ Button.isSmall ] [ str "Small" ]
+    Button.button_a [ ] [ str "Normal" ]
+    Button.button_a [ Button.isMedium ] [ str "Medium" ]
+    Button.button_a [ Button.isLarge ] [ str "Large" ]
 ```
     """
 
@@ -34,32 +34,32 @@ let sizeCode =
 let outlinedCode =
     """
 ```fsharp
-    Button.button_div [ Button.isOutlined ] [ str "Outlined" ]
-    Button.button_div [ Button.isSuccess; Button.isOutlined ] [ str "Outlined" ]
-    Button.button_div [ Button.isPrimary; Button.isOutlined ] [ str "Outlined" ]
-    Button.button_div [ Button.isInfo; Button.isOutlined ] [ str "Outlined" ]
-    Button.button_div [ Button.isDanger;  Button.isOutlined ] [ str "Outlined" ]
+    Button.button_a [ Button.isOutlined ] [ str "Outlined" ]
+    Button.button_a [ Button.isSuccess; Button.isOutlined ] [ str "Outlined" ]
+    Button.button_a [ Button.isPrimary; Button.isOutlined ] [ str "Outlined" ]
+    Button.button_a [ Button.isInfo; Button.isOutlined ] [ str "Outlined" ]
+    Button.button_a [ Button.isDanger;  Button.isOutlined ] [ str "Outlined" ]
 ```
     """
 
 let mixedStyleCode =
     """
 ```fsharp
-    Button.button_div [ Button.isInverted ] [ str "Inverted" ]
-    Button.button_div [ Button.isSuccess; Button.isInverted ] [ str "Inverted" ]
-    Button.button_div [ Button.isDanger; Button.isInverted; Button.isOutlined ] [ str "Invert Outlined" ]
-    Button.button_div [ Button.isInfo; Button.isInverted; Button.isOutlined ] [ str "Invert Outlined" ]
+    Button.button_a [ Button.isInverted ] [ str "Inverted" ]
+    Button.button_a [ Button.isSuccess; Button.isInverted ] [ str "Inverted" ]
+    Button.button_a [ Button.isDanger; Button.isInverted; Button.isOutlined ] [ str "Invert Outlined" ]
+    Button.button_a [ Button.isInfo; Button.isInverted; Button.isOutlined ] [ str "Invert Outlined" ]
 ```
     """
 
 let stateCode =
     """
 ```fsharp
-    Button.button_div [ ] [ str "Normal" ]
-    Button.button_div [ Button.isSuccess; Button.isHovered ] [ str "Hover" ]
-    Button.button_div [ Button.isWarning; Button.isFocused ] [ str "Focus" ]
-    Button.button_div [ Button.isInfo; Button.isActive ] [ str "Active" ]
-    Button.button_div [ Button.isBlack; Button.isLoading ] [ str "Loading" ]
+    Button.button_a [ ] [ str "Normal" ]
+    Button.button_a [ Button.isSuccess; Button.isHovered ] [ str "Hover" ]
+    Button.button_a [ Button.isWarning; Button.isFocused ] [ str "Focus" ]
+    Button.button_a [ Button.isInfo; Button.isActive ] [ str "Active" ]
+    Button.button_a [ Button.isBlack; Button.isLoading ] [ str "Loading" ]
 ```
     """
 
@@ -67,22 +67,22 @@ let extraCode =
     """
 ```fsharp
     // For registering a click event, we can use the Button.onClick helper
-    Button.button_div [ Button.onClick (fun _ -> dispatch Click) ]
-                      [ str buttonTxt ]
+    Button.button_a [ Button.onClick (fun _ -> dispatch Click) ]
+                    [ str buttonTxt ]
     // Or we can pass any IProps via Button.props
     // Equivalent of the Button.onClick
-    Button.button_div [ Button.props [ OnClick (fun _ -> dispatch Click) ] ]
-                      [ str buttonTxt ]
+    Button.button_a [ Button.props [ OnClick (fun _ -> dispatch Click) ] ]
+                    [ str buttonTxt ]
     // Disabled button
-    Button.button_div [ Button.props [ Disabled true ] ]
-                      [ str "Fixed width" ]
+    Button.button_a [ Button.props [ Disabled true ] ]
+                    [ str "Fixed width" ]
 ```
     """
 
 let staticView =
     """
 ```fsharp
-    Button.button_div [ Button.isStatic ]
+    Button.button_a [ Button.isStatic ]
         [ str "Static" ]
 ```
     """
@@ -90,29 +90,41 @@ let staticView =
 let disabled =
     """
 ```fsharp
-    Button.button_div [ Button.isDisabled
-                        Button.isLink ] [ str "Link" ]
-    Button.button_div [ Button.isDisabled
-                        Button.isPrimary ] [ str "Primary" ]
-    Button.button_div [ Button.isDisabled
-                        Button.isInfo ] [ str "Info" ]
-    Button.button_div [ Button.isDisabled
-                        Button.isSuccess ] [ str "Success" ]
-    Button.button_div [ Button.isDisabled
-                        Button.isWarning ] [ str "Warning" ]
-    Button.button_div [ Button.isDisabled
-                        Button.isDanger ] [ str "Danger" ]
+    Button.button_a [ Button.isDisabled
+                      Button.isLink ] [ str "Link" ]
+    Button.button_a [ Button.isDisabled
+                      Button.isPrimary ] [ str "Primary" ]
+    Button.button_a [ Button.isDisabled
+                      Button.isInfo ] [ str "Info" ]
+    Button.button_a [ Button.isDisabled
+                      Button.isSuccess ] [ str "Success" ]
+    Button.button_a [ Button.isDisabled
+                      Button.isWarning ] [ str "Warning" ]
+    Button.button_a [ Button.isDisabled
+                      Button.isDanger ] [ str "Danger" ]
 ```
     """
 
 let icons =
     """
 ```fsharp
-    Button.button_div [ ] [ Icon.faIcon [ ] [ Fa.icon Fa.I.Bold ] ]
-    Button.button_div [ ] [ Icon.faIcon [ ] [ Fa.icon Fa.I.Italic ] ]
-    Button.button_div [ ] [ Icon.faIcon [ ] [ Fa.icon Fa.I.Underline ] ]
-    Button.button_div [ Button.isDanger
-                        Button.isOutlined ] [ str "Danger" ]
+    Button.button_a [ ] [ Icon.faIcon [ ] [ Fa.icon Fa.I.Bold ] ]
+    Button.button_a [ ] [ Icon.faIcon [ ] [ Fa.icon Fa.I.Italic ] ]
+    Button.button_a [ ] [ Icon.faIcon [ ] [ Fa.icon Fa.I.Underline ] ]
+    Button.button_a [ Button.isDanger
+                      Button.isOutlined ] [ str "Danger" ]
+```
+    """
+
+let demoHelpers =
+    """
+```fsharp
+    Button.button_a [ ]
+      [ str "I am an anchor button"]
+    Button.button_btn [ ]
+      [ str "I am a form button"]
+    Button.button_input [ Button.value "I am an input button"
+                          Button.typeIsReset ]
 ```
     """
 
@@ -126,6 +138,7 @@ The **buttons** can have different colors, sizes and states.
 *[Bulma documentation](http://bulma.io/documentation/elements/button/)*
         """
       ColorViewer = Viewer.State.init colorCode
+      DemoHelpersViewer = Viewer.State.init demoHelpers
       SizeViewer = Viewer.State.init sizeCode
       OutlinedViewer = Viewer.State.init outlinedCode
       MixedStyleViewer = Viewer.State.init mixedStyleCode
@@ -141,6 +154,10 @@ let update msg model =
     | ColorViewerMsg msg ->
         let (viewer, viewerMsg) = Viewer.State.update msg model.ColorViewer
         { model with ColorViewer = viewer }, Cmd.map ColorViewerMsg viewerMsg
+
+    | DemoHelpersViewerMsg msg ->
+        let (viewer, viewerMsg) = Viewer.State.update msg model.DemoHelpersViewer
+        { model with DemoHelpersViewer = viewer }, Cmd.map DemoHelpersViewerMsg viewerMsg
 
     | SizeViewerMsg msg ->
         let (viewer, viewerMsg) = Viewer.State.update msg model.SizeViewer