+++ docs/src/Fulma/Elements/Button/Types.fs
@@ -3,6 +3,7 @@ module Elements.Button.Types
 type Model =
     { Intro : string
       ColorViewer : Viewer.Types.Model
+      DemoHelpersViewer : Viewer.Types.Model
       SizeViewer : Viewer.Types.Model
       OutlinedViewer : Viewer.Types.Model
       MixedStyleViewer : Viewer.Types.Model
@@ -15,6 +16,7 @@ type Model =
 
 type Msg =
     | ColorViewerMsg of Viewer.Types.Msg
+    | DemoHelpersViewerMsg of Viewer.Types.Msg
     | SizeViewerMsg of Viewer.Types.Msg
     | OutlinedViewerMsg of Viewer.Types.Msg
     | MixedStyleViewerMsg of Viewer.Types.Msg