+++ docs/src/Fulma/Layout/Level/View.fs
@@ -22,7 +22,7 @@ let iconInteractive =
                         [ Input.input [ Input.typeIsText
                                         Input.placeholder "Find a post" ] ]
                       Control.control_div [ ]
-                        [ Button.button_div [ ]
+                        [ Button.button_a [ ]
                             [ str "Search" ] ] ] ] ]
           Level.right [ ]
             [ Level.item [ ]
@@ -34,7 +34,7 @@ let iconInteractive =
               Level.item [ ]
                 [ a [ ] [ str "Deleted" ] ]
               Level.item [ ]
-                [ Button.button_div [ Button.isSuccess ] [ str "New" ] ] ] ]
+                [ Button.button_a [ Button.isSuccess ] [ str "New" ] ] ] ]
 
 let centered =
     Level.level [ ]