+++ src/Fulma/Elements/Button.fs
@@ -25,6 +25,15 @@ module Button =
             | IsStatic
             | Nothing
 
+        type IAnchorOption =
+            interface end
+
+        type IBtnOption =
+            interface end
+
+        type IInputOption =
+            interface end
+
         type Option =
             | Level of ILevelAndColor
             | Size of ISize
@@ -36,6 +45,18 @@ module Button =
             | Props of IHTMLProp list
             | OnClick of (MouseEvent -> unit)
             | CustomClass of string
+            interface IAnchorOption
+            interface IBtnOption
+            interface IInputOption
+
+        type AnchorOnlyOption =
+            | Href of string
+            interface IAnchorOption
+
+        type InputOnlyOption =
+            | Type of string
+            | Value of string
+            interface IInputOption
 
         let ofSize size =
             match size with
@@ -84,6 +105,38 @@ module Button =
                   CustomClass = None
                   OnClick = None }
 
+        type AnchorOptions =
+            { GenericOption : Options
+              Href : string option }
+
+            static member Empty =
+                { GenericOption = Options.Empty
+                  Href = None }
+
+        type InputOptions =
+            { GenericOption : Options
+              Value : string option
+              Type : string option }
+
+            static member Empty =
+                { GenericOption = Options.Empty
+                  Type = None
+                  Value = None }
+
+
+        let parseGenericOptions (result: Options) opt =
+            match opt with
+            | Option.Level level -> { result with Level = ofLevelAndColor level |> Some }
+            | Size size -> { result with Size = ofSize size |> Some }
+            | IsOutlined -> { result with IsOutlined = true }
+            | IsInverted -> { result with IsInverted = true }
+            | IsLink -> { result with IsLink = true }
+            | State state -> { result with State = ofState state |> Some }
+            | Props props -> { result with Props = props }
+            | CustomClass customClass -> { result with CustomClass = Some customClass }
+            | OnClick cb -> { result with OnClick = cb |> Some }
+            | IsDisabled -> { result with IsDisabled = true }
+
     open Types
 
     // Sizes
@@ -117,38 +170,72 @@ module Button =
     let customClass = CustomClass
     let onClick cb = OnClick cb
 
-    let internal button (element : IHTMLProp list -> ReactElement list -> ReactElement) (options : Option list) children =
-        let parseOptions (result: Options) opt =
+    // Anchor only
+    let href = AnchorOnlyOption.Href
+
+    // Input only
+    let typeIsSubmit = InputOnlyOption.Type "reset"
+    let typeIsReset = InputOnlyOption.Type "submit"
+    let value = InputOnlyOption.Value
+
+    let genericPropsGenerator opts =
+        [ yield classBaseList Bulma.Button.Container
+             [ Bulma.Button.Styles.IsOutlined, opts.IsOutlined
+               Bulma.Button.Styles.IsInverted, opts.IsInverted
+               Bulma.Button.Styles.IsLink, opts.IsLink
+               opts.Level.Value, opts.Level.IsSome
+               opts.Size.Value, opts.Size.IsSome
+               opts.State.Value, opts.State.IsSome
+               opts.CustomClass.Value, opts.CustomClass.IsSome ] :> IHTMLProp
+          if opts.IsDisabled then
+            yield Disabled true :> IHTMLProp
+          if opts.OnClick.IsSome then
+            yield DOMAttr.OnClick opts.OnClick.Value :> IHTMLProp
+          yield! opts.Props ]
+
+    let button_a (options : IAnchorOption list) children =
+        let parseOptions (result: AnchorOptions) (opt : IAnchorOption) =
             match opt with
-            | Option.Level level -> { result with Level = ofLevelAndColor level |> Some }
-            | Size size -> { result with Size = ofSize size |> Some }
-            | IsOutlined -> { result with IsOutlined = true }
-            | IsInverted -> { result with IsInverted = true }
-            | IsLink -> { result with IsLink = true }
-            | State state -> { result with State = ofState state |> Some }
-            | Props props -> { result with Props = props }
-            | CustomClass customClass -> { result with CustomClass = Some customClass }
-            | OnClick cb -> { result with OnClick = cb |> Some }
-            | IsDisabled -> { result with IsDisabled = true }
+            | :? Option as genericOption -> { result with GenericOption = parseGenericOptions result.GenericOption genericOption }
+            | :? AnchorOnlyOption as anchorOption ->
+                match anchorOption with
+                | AnchorOnlyOption.Href href -> { result with Href = Some href }
+            | invalidType ->
+                Fable.Import.JS.console.warn ("Invalid option type given for the anchor button: " + string invalidType)
+                result
 
-        let opts = options |> List.fold parseOptions Options.Empty
+        let anchorOpts = options |> List.fold parseOptions AnchorOptions.Empty
 
-        element
-            [ yield classBaseList
-                (Helpers.generateClassName Bulma.Button.Container [ opts.Level; opts.Size; opts.State ])
-                 [ Bulma.Button.Styles.IsOutlined, opts.IsOutlined
-                   Bulma.Button.Styles.IsInverted, opts.IsInverted
-                   Bulma.Button.Styles.IsLink, opts.IsLink
-                   opts.CustomClass.Value, opts.CustomClass.IsSome ] :> IHTMLProp
-              if opts.IsDisabled then
-                yield Disabled true :> IHTMLProp
-              if opts.OnClick.IsSome then
-                yield DOMAttr.OnClick opts.OnClick.Value :> IHTMLProp
-              yield! opts.Props ]
+        a
+            [ if anchorOpts.Href.IsSome then
+                yield HTMLAttr.Href anchorOpts.Href.Value :> IHTMLProp
+              yield! genericPropsGenerator anchorOpts.GenericOption ]
             children
 
-    let button_a (options : Option list) children =
-        button a options children
+    let button_btn (options : Option list) children =
+        let opts = options |> List.fold parseGenericOptions Options.Empty
 
-    let button_div (options : Option list) children =
-        button div options children
+        button
+            (genericPropsGenerator opts)
+            children
+
+    let button_input (options : IInputOption list) =
+        let parseOptions (result: InputOptions) (opt : IInputOption) =
+            match opt with
+            | :? Option as genericOption -> { result with GenericOption = parseGenericOptions result.GenericOption genericOption }
+            | :? InputOnlyOption as anchorOption ->
+                match anchorOption with
+                | InputOnlyOption.Type typeValue -> { result with Type = Some typeValue }
+                | InputOnlyOption.Value value -> { result with Value = Some value }
+            | invalidType ->
+                Fable.Import.JS.console.warn ("Invalid option type given for the input button: " + string invalidType)
+                result
+
+        let inputOpts = options |> List.fold parseOptions InputOptions.Empty
+
+        input
+            [ if inputOpts.Type.IsSome then
+                yield HTMLAttr.Type inputOpts.Type.Value :> IHTMLProp
+              if inputOpts.Value.IsSome then
+                yield HTMLAttr.Value inputOpts.Value.Value :> IHTMLProp
+              yield! genericPropsGenerator inputOpts.GenericOption ]